app/pyproject.toml
[project]
name = "test-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "baml-py>=0.218.1",
    "boto3>=1.42.43",
    "fastapi>=0.128.2",
    "httpx>=0.28.1",
    "langfuse>=3.12.1",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "ruff>=0.15.0",
    "uvicorn>=0.40.0",
]
